ref: change data types

+++ scalar/src/lib.rs
@@ -10,10 +10,10 @@ pub fn pro(a: i8, b: i8) -> i8 {
     a * b
 }
 
-pub fn quo(a: i32, b: i32) -> i32 {
+pub fn quo(a: f64, b: f64) -> f64 {
         a / b
 }
 
-pub fn rem(a: i32, b: i32) -> i32 {
+pub fn rem(a: f64, b: f64) -> f64 {
         a % b
 }
